+++ src/lib/utils/gas-estimatooor.ts
@@ -27,7 +27,7 @@ export class GasEstimatooor {
     try {
       gasEstimate = await signer.estimateGas(tx)
     } catch (error: any) {
-      console.log('Error estimating gas:', error)
+      // console.log('Error estimating gas:', error)
       if (canFail) {
         throw new GasEstimatooorFailedError()
       }
